fix run_async_function_within_sync_function
