+++ litellm/litellm_core_utils/asyncify.py
@@ -1,5 +1,6 @@
+import asyncio
 import functools
-from typing import Awaitable, Callable, Optional
+from typing import Any, Awaitable, Callable, Coroutine, Optional
 
 import anyio
 import anyio.to_thread
@@ -67,3 +68,13 @@ def asyncify(
         )
 
     return wrapper
+
+
+def run_async_function_within_sync_function(coroutine: Coroutine) -> Any:
+    """
+    Helper to run an async function from a sync function using anyio.
+
+    Usage:
+        result = run_async_function(my_async_func(arg1, arg2))
+    """
+    return anyio.run(lambda: coroutine)
